feat: todo 기능 추가
todo 등록, done checkbox 업데이트 기능 추가

+++ src/main/java/com/dylc/DYLC/repository/TodoRepository.java
@@ -1,8 +1,11 @@
 package com.dylc.DYLC.repository;
 
 import com.dylc.DYLC.entity.Todo;
+import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
-public interface TodoRepository extends JpaRepository<Todo, Integer> {
+
+public interface TodoRepository extends JpaRepository<Todo, Long> {
 
 }
